Use internal solver
Uses @mauro3's PR for an example solver to build an internal ODE solver that can be used with tests. This decouples the Base library tests from all downstream packages, making it independent of outside changes. This means that it can no longer check whether it breaks other packages, which is a better job for JuliaCI anyways. But it also means that DiffEqBase is safe when upper bounds on downstream packages are placed due to interface updates. 

@tkelman pinging you so you know about this update. Hopefully it makes your life easier. I wish I thought of this sooner.