Support for immutable fields derived from AbstractArray
The interface provided by ArrayInterface.jl is now used to check
whether the fields of the structure derived from `DEDataArray` are
immutable. Hence, we can now have fields derived from
`AbstractArray` that are immutable.

Notice that it was required to move the file with the related tests
to the beginning of the set. Since `recursivecopy!` and
`copy_fields!` are both generated functions, then the structure
`Quaternion`, which was created specifically for this test,
together with the definition of the method
`ArrayInterface.ismutable(::Type{<:Quaternion})` must be placed
before the first call to `using DiffEqBase` in the current Julia
session.

Closes #507