make dual norms include the dual terms
This changes the norms on dual numbers to include the dual portions. This means that the solution to the ODE under automatic differentiation is not necessarily the same as without. However, https://github.com/SciML/DiffEqSensitivity.jl/issues/273 nicely highlights some edge cases that show that this is somewhat necessary. In that case, the integrator has zero error, and so time steps are unaffected by tolerances which causes gradients to explode since they are not tolerance controlled. This change adds them to the summation and the length, making the solution by dual numbers equivalent to the result of forward sensitivity analysis under the default norm.

@yingboma what do you think?

@pkofod you might want to test this with Pumas. It should give it some more robustness.