mass matrix I for DynamicalODEFunction
You know how sometimes you later think that some idea was a really really bad idea? Yes, this was one of them. "It's a pair of ODEs, so it's a pair of mass matrices!". Yeah, but that breaks all implicit solver fallbacks, and even when implicit methods specialized to second order ODEs exist, mass_matrix=I will still work as the special case to catch, so it's just better...

Anyways, it didn't matter until last week when ArrayPartition suddenly became compatible with all implicit solvers, except for the fact that it would fail because the mass matrix was by default a tuple. So, let's fix this and yay, 2nd order ODEs are now better.