More robust method checking before wrapping
Fixes https://github.com/SciML/DiffEqBase.jl/issues/821 . The key is that promote_type ends up calling promote_rule, so it always has a method. We need to check if the promotion rules exist with duals to know if this we can do the wrapping.

In the future, we can do further customization of the wrappers based on if the solver does AD or not, though that would likely increase compile times by a lot (since then there would be multiple versions of the solver), so I'm not convinced that's the right approach after. At least this is always safe, if not "too safe" in some sense.