Add function and parameterization properties to `NullODEIntegrator`
The `NullODEIntegrator` was missing a few pieces needed to satisfy the
implicit contract required to index into it with observables.  This
closes the gap, storing the problem's `.f` and `.p` fields.

It might make sense to eventually re-work these APIs to use accessor
functions such as `get_func()` or `get_parameterization()`, which can
then flexibly thread through to `.prob.f` or similar, although I
understand such a rework is not exactly appealing work.