Fix downstream AD tests
This is because the norm on (u=tracked, t=tracked) returns a tracked value, which is required for differentiating appropriately w.r.t. callback time. However, this value is just being saved for non-differentiable logic in order to know whether to include the first interval or not in the search process, and so it should be dropped.